+++ Robust.Shared/Toolshed/ToolshedCommand.Help.cs
@@ -1,5 +1,6 @@
-﻿using System.Linq;
-using Robust.Shared.Localization;
+﻿using System;
+using System.Linq;
+using System.Text;
 
 namespace Robust.Shared.Toolshed;
 
@@ -31,15 +32,65 @@ public abstract partial class ToolshedCommand
     /// </summary>
     public string GetHelp(string? subCommand)
     {
-        if (subCommand is null)
-            return $"{Name}: {Description(null)}";
-        else
-            return $"{Name}:{subCommand}: {Description(subCommand)}";
+        // Description
+        var description = subCommand is null
+            ? $"{Name}: {Description(null)}"
+            : $"{Name}:{subCommand}: {Description(subCommand)}";
+
+        // Usage
+        var usage = new StringBuilder();
+        usage.AppendLine();
+        usage.Append(Loc.GetString("command-description-usage"));
+        foreach (var (pipedType, parameters) in _readonlyParameters[subCommand ?? ""])
+        {
+            usage.Append(Environment.NewLine + "  ");
+
+            // Piped type
+            if (pipedType != null)
+            {
+                usage.Append(Loc.GetString("command-description-usage-pipedtype",
+                    ("typeName", GetFriendlyName(pipedType))));
+            }
+
+            // Name
+            usage.Append(Name);
+
+            // Parameters
+            foreach (var param in parameters)
+            {
+                usage.Append($" <{GetFriendlyName(param)}>");
+            }
+        }
+
+        return description + usage;
     }
 
     /// <inheritdoc/>
     public override string ToString()
     {
         return GetHelp(null);
+    }
+
+    public static string GetFriendlyName(Type type)
+    {
+        string friendlyName = type.Name;
+        if (type.IsGenericType)
+        {
+            int iBacktick = friendlyName.IndexOf('`');
+            if (iBacktick > 0)
+            {
+                friendlyName = friendlyName.Remove(iBacktick);
+            }
+            friendlyName += "<";
+            Type[] typeParameters = type.GetGenericArguments();
+            for (int i = 0; i < typeParameters.Length; ++i)
+            {
+                string typeParamName = GetFriendlyName(typeParameters[i]);
+                friendlyName += (i == 0 ? typeParamName : "," + typeParamName);
+            }
+            friendlyName += ">";
+        }
+
+        return friendlyName;
     }
 }